+++ src/select.js
@@ -147,6 +147,16 @@ define(function(require, exports, module) {
             var oldSelectIndex = this.get('selectedIndex');
             this.set('selectedIndex', selectIndex);
 
+            // 同步 html 到 model
+            var model = this.get('model');
+            if (oldSelectIndex >= 0) {
+                model.select[oldSelectIndex].selected = false;
+            }
+            if (selectIndex >= 0) {
+                model.select[selectIndex].selected = true;
+            }
+            this.set('model', model);
+
             // 如果不是原来选中的则触发 change 事件
             if (oldSelectIndex !== selectIndex) {
                 var current = this.options.eq(selectIndex);
@@ -214,6 +224,7 @@ define(function(require, exports, module) {
         enableOption: function(option) {
             var index = getOptionIndex(option, this.options);
             var model = this.get("model").select;
+            console.log(model);
             model[index].disabled = false;
             this.syncModel(model);
             return this;