+++ src/select.js
@@ -108,10 +108,10 @@ define(function(require, exports, module) {
                 // 如果 name 存在则创建隐藏域
                 selectName = this.get('name');
                 if (selectName) {
-                    var input = $('input[name=' + selectName + ']').eq(0);
+                    var input = $('input[name="' + selectName + '"]').eq(0);
                     if (!input[0]) {
                         input = $(
-                            '<input type="hidden" id="select-' + selectName +
+                            '<input type="hidden" id="select-' + selectName.replace(/\./g, '-') +
                             '" name="' + selectName +
                             '" />'
                         ).insertAfter(trigger);