+++ HISTORY.md
@@ -5,6 +5,7 @@
 ## 0.9.8
 
 `tag:improved` 升级依赖到 overlay@1.1.4 和 template@0.9.2 。
+`tag:new` 添加 enableOption 和 disableOption 方法。
 
 ## 0.9.7